pyproject.toml
[project]
name = "CannotMax-Greenvine"
version = "1.0.8"
description = "这是一个基于深度学习的明日方舟游戏辅助工具，用于自动识别游戏画面中的单位并预测战斗结果。"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib",
  "numpy",
  "onnxruntime",
  "opencv-python",
  "pandas",
  "pillow",
  "pyqt6",
  "rapidocr",
  "scikit-learn",
  "windows-capture",
  "toml",
  "pywin32>=311",
  "maafw>=5.10.2",
]

[dependency-groups]
dev = ["onnx>=1.19.0", "onnxscript>=0.7.0", "pyinstaller>=6.15.0"]

[project.optional-dependencies]
cpu = ["torch", "torchvision"]
cu128 = ["torch>=2.7.0", "torchvision>=0.22.0"]
cu130 = ["torch>=2.7.0", "torchvision>=0.22.0"]

[tool.uv]
python-install-mirror = "https://mirror.nju.edu.cn/github-release/astral-sh/python-build-standalone/"
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu128" },
  ],
  [
    { extra = "cpu" },
    { extra = "cu130" },
  ],
  [
    { extra = "cu128" },
    { extra = "cu130" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
  { index = "pytorch-cu130", extra = "cu130" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
  { index = "pytorch-cu130", extra = "cu130" },
]

[[tool.uv.index]]
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
default = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
